compact symbol rendering into the common interface for better locality...
